requirements.txt
flask
fastapi
uvicorn[standard]
gunicorn
pyyaml
pyinstaller
